refactor(local-service): split execution prompt helpers
